Support multiple for media selector
